Converted indices to 16 bit

+++ App/include/MeshData.h
@@ -13,7 +13,7 @@ struct Vertex {
 
 struct MeshData {
 	std::vector<Vertex> vertices;
-	std::vector<uint32_t> indices;
+	std::vector<uint16_t> indices;
 	std::vector<std::string> textures;
 };
